### 2018-03-07 #608 @DTChuck
The hyperlinks in the generated PDF are not blue. Is that expected?
