### 2018-03-09 #608 @arfon
> The hyperlinks in the generated PDF are not blue. Is that expected?

@DTChuck yes, I think that's normal.

Would you mind taking a look at the [list of JOSS reviewers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) and seeing if you think there are any that could be good reviewers here?
