### 2018-03-15 #630 @whedon
[REVIEW]: HSImage: A Python and C++ library to allow interaction with ENVI-BIL hyperspectral images
**Submitting author:** @DTChuck (<a href="http://orcid.org/0000-0002-3832-775X">Ryan Brown</a>)
**Repository:** <a href="https://github.com/DTChuck/HSImage" target ="_blank">https://github.com/DTChuck/HSImage</a>
**Version:** v1.2.0
