### 2018-04-10 #630 @abhishekbajpayee
@DTChuck here are some questions I have:
1. Does this require Python 3.x and won't work for Python 2.x? If so, why is that? I ask because even though a lot of people are moving to Python 3.x, adoption might suffer because of the Python 3.x requirement because a ton of people still use Python 2.x. If they end up having to install Python 3.x then managing 2 versions of Python becomes confusing.
4. I tried to create a `virtualenv` and then tried to install HSI using `pip install HSI` but that did not seem to work. However, installing works if I just run `pip install HSI` outside a `virtualenv` Any idea why?

Here is a checklist of suggested fixes:
- [ ] Clearly state, both in documentation and in the repo readme installation instructions, what the dependencies are. Ideally, include specific commands that a user can run in order to install the dependencies.
- [ ] Links to images and documentation in the repo readme are broken. The text is correct but links to the wrong destinations.
