### 2018-04-11 #630 @DTChuck
Addressing 2: I haven't seen that issue before. I was able to install it in a virtualenv on both my Mac and Ubuntu PC, but it could just be some setting I have. It could possibly be due to the OpenCV installation not being present in the virtualenv, as that can't be installed via pip as a dependency. If you have any logs or error messages, I can take a look and see what may be happening.

As for 1: All of my development environment is Python3. I will look to see if it is feasible to use Python2 as well.

What did you find confusing about the installation instructions? I thought that I included all of the dependencies in the installation instructions. I can take another look at that and get it fixed up. I will work on the links and get those fixed as well. Thanks for your feedback!
