### 2018-04-11 #630 @abhishekbajpayee
Regarding installation instructions, I think it would be better if you list the dependencies as a list. Also, it would make it super easy for others if you include a few lines that users can just copy and paste into a terminal to install the dependencies. I am pretty sure it is possible to install Boost libs using `apt-get`. Installing OpenCV can be a bit tricky but I know that it is possible to install OpenCV in one line using `conda`.

I doubt that the issue with installing in a `virtualenv` is related to OpenCV. When I try to install in a `virtualenv`, I see the following:
```
ab9@ab9-laptop:~/projects/HSI$ source bin/activate
(HSI) ab9@ab9-laptop:~/projects/HSI$ pip install HSI
Collecting HSI
  Could not find a version that satisfies the requirement HSI (from versions: )
No matching distribution found for HSI
(HSI) ab9@ab9-laptop:~/projects/HSI$
```
