### 2018-04-11 #630 @DTChuck
I am going to update the instructions later today. I am also spooling up an Ubuntu VM so I can test python 2.7 and 3.x installation in a clean environment. I will check both of the boxes above once I have finished. Thanks for the feedback!
