+++ README.md
@@ -3,7 +3,7 @@
 Interface Library for ENVI-BIL Hyperspectral Images
 (C++/Python)
 
-This library is designed to allow open-source experimentation with ENVI-BIL hyperspectral images. The goal of this software is to promote the study of hyperspectral images in the academic computer vision research world. This software provides both a C++ and Python interface for ease of use. An associated project has resulted in a publicly available library of images available at [https://osf.io/pd49t/](). These images are all compatible and intended to be used with this software interface library.
+This library is designed to allow open-source experimentation with ENVI-BIL hyperspectral images. The goal of this software is to promote the study of hyperspectral images in the academic computer vision research world. This software provides both a C++ and Python interface for ease of use. An associated project has resulted in a publicly available library of images available at [https://osf.io/pd49t/]([https://osf.io/pd49t/). These images are all compatible and intended to be used with this software interface library.
 
 ## Installation ##
 This package requires the boost crossplatfrom C++ libraries and OpenCV 3.x for the C++ library. To use the package in Python 3.x, OpenCV 3.x is required as a Python module.
@@ -26,7 +26,7 @@ Cloning the repository will include the test suite which can be run using setupt
 This test **REQUIRES** external data (hyperspectral image files) to run. These files are stored on GitHub using the Git LFS protocol. Cloning the repository should automatically download the files to the appropriate folder. 
 
 ## Documentation ## 
-Full documentation is available at [https://dtchuck.github.io/HSImage/]()
+Full documentation is available at [https://dtchuck.github.io/HSImage/]([https://dtchuck.github.io/HSImage/)
 
 ## Example Usage ##
 Loading a image, converting to a pixel vector array and normalizing by the ambient intensity to create spectral reflectance data.
