### 2018-04-11 21a8855
Added install_opencv.sh
