### 2018-04-11 661f70b
Updated README and Doxygen docs
