### 2018-04-11 #630 @DTChuck
I have checked everything over. I added a list of dependencies and tested installation on Python 2. It looks like it installs and tests just fine on Python 2.7 so I added that to the dependency list. I fixed the dead links in the README file. Let me know if you want me to take a look at anything else repo related.

As far as your installation error, I still haven't been able to reproduce it. I tested both Python 2.7 and 3.4 in new virtualenvs on a clean VM today, and it installed just fine. Was the virtualenv a Python 2 or Python 3 environment? Its possible that it just did not build because the setup.py file did not allow Python 2 builds on the version you were testing. That should be different now. If you still have this error, we can look into some more debugging steps. Thanks!!
