### 2018-05-03 #630 @abhishekbajpayee
Next comments from my end:
1. I am not sure how to check for release version because I don't see any GitHub releases in the project repo. @arfon should I be looking somewhere else?
2. Even after running `pip install virtualenvwrapper`, I don't see any script called `virtualenvwrapper.sh` in `/usr/local/bin`. I checked the official documentation for `virtualenvwrapper` at http://virtualenvwrapper.readthedocs.io/en/latest/index.html and they seem to suggest the same steps as you list. For some reason, it is not working for me. Is it necessary to use `virtualenvwrapper`? If not, it would be best not to mention this in the installation steps because it is likely to confuse others too.
On Ubuntu 16.04, given all other dependencies are satisfied, I was able to install by just using traditional `virtualenv` by running the following (my guess is that this should work on most other versions of Ubuntu as well):
```
$ virtualenv hsi_env
$ cd hsi_env
$ source bin/activate
$ pip install HSI
```
3. Creating the symlinks for `cv2.so` wasn't required for me to successfully import `HSI`.
4. Since you do require the python bindings for OpenCV to be built, it would be useful to add a short note saying if someone is installing OpenCV 2.4.x then they should ensure building of python bindings is enabled by running `cmake -DWITH_PYTHON=ON ..` in the build folder.
5. The URL for full documentation listed under the "Documentation" does not link to the URL. Would be ideal to link. Same goes for URL for library of images.
6. I am able to import `HSI` in a python shell but wasn't able to test on any data. Mainly because I am not sure how to clone the library of images. Would be good to include instructions on how to do that. Also, how large is this dataset?

Not critical but things to think about:
1. Overall, the `install_opencv.sh` script is greatly and will be useful for a bunch of folks. I see that it is explicitly installing `python3-dev`, `python3-tk` and `python3-numpy`. Is this intended / necessary?
2. In the `install_opencv.sh` script, are the flags passed to `cmake` required? It seems that a simple `cmake ..` configures the build just fine.

Sorry about the delayed response. I will be sure to set aside time every couple days to respond so we can get this done soon.
