+++ README.md
@@ -3,7 +3,7 @@
 Interface Library for ENVI-BIL Hyperspectral Images
 (C++/Python)
 
-This library is designed to allow open-source experimentation with ENVI-BIL hyperspectral images. The goal of this software is to promote the study of hyperspectral images in the academic computer vision research world. This software provides both a C++ and Python interface for ease of use. An associated project has resulted in a publicly available library of images available at [https://osf.io/pd49t/]([https://osf.io/pd49t/). These images are all compatible and intended to be used with this software interface library.
+This library is designed to allow open-source experimentation with ENVI-BIL hyperspectral images. The goal of this software is to promote the study of hyperspectral images in the academic computer vision research world. This software provides both a C++ and Python interface for ease of use. An associated project has resulted in a publicly available library of images available at [https://osf.io/pd49t/](https://osf.io/pd49t/). These images are all compatible and intended to be used with this software interface library.
 
 ## Installation ##
 NOTE: This software, while it is build with only cross-platform libraries, has not been tested on Windows. It currently has been tested on MacOS 10.11.x and Ubuntu 14.04. If a user wishes to use this software on Windows, it should be possible to build and run, but there may be issues in the linking of the OpenCV libraries. Environment variable settings on Windows can be tricky, and there is no official Windows support for HSImage at this time.
@@ -70,7 +70,7 @@ Cloning the repository will include the test suite which can be run using setupt
 This test **REQUIRES** external data (hyperspectral image files) to run. These files are stored on GitHub using the Git LFS protocol. Cloning the repository should automatically download the files to the appropriate folder. 
 
 ## Documentation ## 
-Full documentation is available [here]([https://dtchuck.github.io/HSImage/)
+Full documentation is available at [https://dtchuck/github.io/HSImage/](https://dtchuck.github.io/HSImage/)
 
 ## Example Usage ##
 Loading a image, converting to a pixel vector array and normalizing by the ambient intensity to create spectral reflectance data.
