### 2018-05-07 #630 @DTChuck
Here are the answers to your questions:
1) I created a release for version 1.2.0. You should be able to see that on the repository page
2) Python best practice is to encapsulate everything in virtual environments. I'm not sure why the `virtualenvwrapper` didn't work for you, but I want to leave virtual environment instructions in the install instructions, as that is the recommended Python workflow.
3) That's good. I know that it might be necessary for some people when creating virtualenvs so I wanted to explicitly mention it.
4) OpenCV 2.x is not supported by this software. This software only supports OpenCV 3.x
5) Fixed the links. Had a bad character in the actual link address in the markdown file, but once i found and removed it, the links worked again.
6) You can test the software by cloning the repository. A single hyperspectral image is included in the GitHub repository and can be cloned along with the repo. You should be able to follow the directions in the testing section of the README.md to test the software.

Additional things:
1) The python3-dev is needed for the compilation of the OpenCV python interface. python3-numpy is needed as OpenCV uses Numpy to store image matrices, and python3-tk is needed to display OpenCV images, as OpenCV Python uses TKinter to display images.
2) The cmake flags are not required, but should provide better performance to most systems. To most users, the differences should be transparent and not cause any issues at all.

Thanks a bunch for getting back to me. Let me know if you need me to look at anything else
