### 2018-05-08 #630 @abhishekbajpayee
Thanks for creating the release version and fixing the links. Next few issues / clarifications:

1. When I was asking do you need to use `virtualenvwrapper`, I specifically meant to ask whether you need to use `virtualenvwrapper` _in addition_ to `virtualenv`. Using only `virtualenv` worked for me and I think that should work for most people. If this is the case, you do not need to mention `virtualenvwrapper` in the install instructions at all.
2. Adding a small note next to "OpenCV 3.x - C++ and Python" saying that OpenCV 2.x is not supported would be very useful. I know a ton of folks still use OpenCV 2.x.
3. You mention that `python3-dev`, `python3-numpy` and `python3-tk` are required for OpenCV functionality. However, it seems that `HSI` works with Python 2. If that is the case, I feel like these packages should not be required. Perhaps only `python-dev`, `python-tk` should be enough and `pip install numpy` should take care of the Numpy requirement?
4. For reading images, I see that 2 files `header_filename` and `raw_filename` are required. Do these files typically have the same filename (just different extensions) and are in the same directory? If so, would it make sense to pass only 1 argument to `HSI.hsimage()` and locate the other file automatically?
5. I think the example section could use more detail and perhaps some images. Can you show some interesting functionality in a visual manner here?
6. When I view the JOSS paper (`paper.md`), I don't see anything under the references section. Is this a bug and needs fixing?

I can close the installation and functionality checks once points 1 through 4 are addressed. That should only leave a couple other checks which I think can be closed by end of this week. Almost there!
