### 2018-05-08 #630 @DTChuck
Glad to hear we are getting close :)

1) I added a separate install instruction for virtualenv on its own. I like virtualenvwrapper as a module as it allows me to simply call `workon <env_name>` to jump into any virtual environment. I left the instructions for virtualenvwrapper in, but as an optional step.

2) Added a note in the README.md and the Documentation stating that OpenCV 2.x is not supported.

3) As far as I know, the python-x libraries are required for Python 2.x support and the python3-x libraries are required for Python 3.x support. I don't know which version of Python the user is going to be running so I figured that it was necessary to install both sets to ensure that it worked for both Python 2 and Python 3. The user SHOULD have installed numpy separately at this point, but I can'
t necessarily trust that they did so before running the script, so I also included a numpy installation so that it is guaranteed to be in place for the OpenCV install

4) I had the same thought earlier in the design process, but I changed to requiring both the names independently because the hope for this software is that it is database agnostic. I would like other people to be able to use it apart from the dataset that I put together, and their file structure and naming conventions may be different than mine. I didn't want to enforce a naming convention on other researchers. That was my reasoning for not automatically picking up the names of the files.

5) I will work on it and hopefully get some more detailed examples up in the next day or so.

6) The references are generated through bibtex and only show up in the compiled PDF. I can see them when I look at the article proof [here](https://github.com/openjournals/joss-papers/blob/joss.00630/joss.00630/10.21105.joss.00630.pdf)
