### 2018-05-08 c27129e
Updated README.md with new install instructions
