### 2018-05-08 3b17e11
Updated README.md with new install instructions
