### 2018-05-09 #630 @DTChuck
I have added an additional example script and placed images in the README file that should display some of the capabilities of the software.
