### 2018-05-14 #630 @DTChuck
I followed the citation information for hsdar as specified on their webpage [here](https://cran.r-project.org/web/packages/hsdar/citation.html). There was no DOI listed.

I followed the citation information for pybind11 as specified [here](https://github.com/pybind/pybind11/commit/e7d304fbc6f0aa22b6105e2b1a4807c7a294eafa). Again, no DOI listed.

The OpenCV citation information also did not contain a DOI. [Citation Info](https://github.com/opencv/opencv/wiki/CiteOpenCV).
