### 2018-05-14 #630 @abhishekbajpayee
@DTChuck last couple things:

1) Is it possible to use matplotlib instead of OpenCV for plotting in Python? If so, I would recommend using that in the example script. I just ran into some issues running the example script because of the `cv.imshow()` calls. In my experience, the OpenCV plotting backend provides very little extra functionality and often requires extra configuration.
2) Can you add something like "Contribute" or "How to contribute" or something similar to the last heading in the README file? That way, anyone wishing to contribute will easily be able to navigate to the right place and figure out that they have to email you.

I will close the References check once I hear from @arfon about the DOI issue.
