### 2018-05-15 #630 @arfon
> @arfon, the papers lists references to OpenCV, pybind11 and hsdar (R package) which do not have DOI numbers. Is that OK?

Yes, that's fine. We only ask for a DOI if there is one available. Thanks for checking!
