### 2018-05-15 544d346
Added how to contribute
