### 2018-05-15 #630 @DTChuck
@abhishekbajpayee
1) I suppose you could use matplotlib for the plotting instead of OpenCV, but the images are stored in OpenCV format and are intended to be used in OpenCV functions. If the imshow() function isn't working, that indicates something has gone wrong in your environment or installation, and other OpenCV functions may also not work properly. It is fairly typical of examples leveraging OpenCV to use the built-in image viewing functions.

2) Added.
