### 2018-05-18 #630 @DTChuck
@abhishekbajpayee
Thanks for all your help on this. I will look into using matplotlib as the base for the example script. I appreciate the work you have done on this review.

@arfon What do I need to do to get this from "under review" to "complete"? Let me know if you need anything from me.
