### 2018-05-21 #630 @DTChuck
Here is the archive DOI. I am also going to place it in the README of the GitHub repository. Thanks!!

[![DOI](https://zenodo.org/badge/96904936.svg)](https://zenodo.org/badge/latestdoi/96904936)
